+++ packages/lib/src/index.ts
@@ -5,7 +5,7 @@ if (process.env.NODE_ENV === 'development') {
     require('preact/debug');
 }
 
-import '../config/polyfills';
+import './polyfills';
 import Checkout from './core';
 /* eslint-enable */
